consumeEventIterator -> consumeAsyncIterator
