+++ packages/client/src/utils.test-d.ts
@@ -1,8 +1,7 @@
-import type { PromiseWithError } from '@orpc/shared'
 import type { ORPCError } from './error'
 import type { Client, ClientContext } from './types'
 import { isInferableError } from './error-utils'
-import { consumeEventIterator, safe } from './utils'
+import { safe } from './utils'
 
 describe('safe', async () => {
   const client = {} as Client<ClientContext, string, number, Error | ORPCError<'BAD_GATEWAY', { val: string }>>
@@ -70,55 +69,5 @@ describe('safe', async () => {
 
     expectTypeOf(error).toEqualTypeOf<Error | null>()
     expectTypeOf(data).toEqualTypeOf<number | undefined>()
-  })
-})
-
-describe('consumeEventIterator', () => {
-  it('can infer types from PromiseWithError + AsyncGenerator', () => {
-    void consumeEventIterator({} as PromiseWithError<AsyncGenerator<'message-value', 'done-value'>, 'error-value'>, {
-      onEvent: (message) => {
-        expectTypeOf(message).toEqualTypeOf<'message-value'>()
-      },
-      onError: (error) => {
-        expectTypeOf(error).toEqualTypeOf<'error-value'>()
-      },
-      onSuccess: (value) => {
-        expectTypeOf(value).toEqualTypeOf<'done-value' | undefined>()
-      },
-      onFinish: ([error, data, isSuccess]) => {
-        if (!error || isSuccess) {
-          expectTypeOf(error).toEqualTypeOf<null>()
-          expectTypeOf(data).toEqualTypeOf<'done-value' | undefined>()
-        }
-        else {
-          expectTypeOf(error).toEqualTypeOf<'error-value'>()
-          expectTypeOf(data).toEqualTypeOf<undefined>()
-        }
-      },
-    })
-  })
-
-  it('can infer types from AsyncIterator', () => {
-    void consumeEventIterator({} as AsyncIterator<'message-value', 'done-value'>, {
-      onEvent: (message) => {
-        expectTypeOf(message).toEqualTypeOf<'message-value'>()
-      },
-      onError: (error) => {
-        expectTypeOf(error).toEqualTypeOf<Error>()
-      },
-      onSuccess: (value) => {
-        expectTypeOf(value).toEqualTypeOf<'done-value' | undefined>()
-      },
-      onFinish: ([error, data, isSuccess]) => {
-        if (!error || isSuccess) {
-          expectTypeOf(error).toEqualTypeOf<null>()
-          expectTypeOf(data).toEqualTypeOf<'done-value' | undefined>()
-        }
-        else {
-          expectTypeOf(error).toEqualTypeOf<Error>()
-          expectTypeOf(data).toEqualTypeOf<undefined>()
-        }
-      },
-    })
   })
 })
