+++ packages/client/src/utils.ts
@@ -61,79 +61,3 @@ export async function safe<TOutput, TError = ThrowableError>(promise: PromiseWit
     )
   }
 }
-
-export interface ConsumeEventIteratorOptions<T, TReturn, TError> {
-  /**
-   * Called on each event
-   */
-  onEvent: (event: T) => void
-  /**
-   * Called once error happens
-   */
-  onError?: (error: TError) => void
-  /**
-   * Called once event iterator is done
-   *
-   * @info If iterator is canceled, `undefined` can be passed on success
-   */
-  onSuccess?: (value: TReturn | undefined) => void
-  /**
-   * Called once after onError or onSuccess
-   *
-   * @info If iterator is canceled, `undefined` can be passed on success
-   */
-  onFinish?: (state: [error: TError, data: undefined, isSuccess: false] | [error: null, data: TReturn | undefined, isSuccess: true]) => void
-}
-
-/**
- * Consumes an event iterator with lifecycle callbacks
- *
- * @warning If no `onError` or `onFinish` is provided, error will be thrown into unhandled rejection channel.
- * @return unsubscribe callback
- */
-export function consumeEventIterator<T, TReturn, TError = ThrowableError>(
-  iterator: AsyncIterator<T, TReturn> | PromiseWithError<AsyncIterator<T, TReturn>, TError>,
-  options: ConsumeEventIteratorOptions<T, TReturn, TError>,
-): () => Promise<void> {
-  void (async () => {
-    let onFinishState: [error: TError, data: undefined, isSuccess: false] | [error: null, data: TReturn | undefined, isSuccess: true]
-
-    try {
-      const resolvedIterator = await iterator
-
-      while (true) {
-        const { done, value } = await resolvedIterator.next()
-
-        if (done) {
-          // if iterator is canceled, value can be undefined
-          const realValue = value as typeof value | undefined
-          onFinishState = [null, realValue, true]
-          options.onSuccess?.(realValue)
-          break
-        }
-
-        options.onEvent(value)
-      }
-    }
-    catch (error) {
-      onFinishState = [error as TError, undefined, false]
-
-      /**
-       * If no `onError` or `onFinish` is provided, unhandled rejections will be thrown
-       * This is best practice for error handling - error should not be silently ignored
-       */
-      if (!options.onError && !options.onFinish) {
-        throw error
-      }
-
-      options.onError?.(error as TError)
-    }
-    finally {
-      options.onFinish?.(onFinishState!)
-    }
-  })()
-
-  return async () => {
-    await (await iterator)?.return?.()
-  }
-}
