+++ packages/shared/src/iterator.ts
@@ -1,5 +1,6 @@
 import type { Promisable } from 'type-fest'
 import type { StartSpanOptions } from './opentelemetry'
+import type { PromiseWithError, ThrowableError } from './types'
 import { AsyncIteratorClass } from '@standardserver/shared'
 import { once } from './function'
 import { recordSpanError, runInSpanContext, startSpan } from './opentelemetry'
@@ -155,4 +156,80 @@ export function replicateAsyncIterator<T, TReturn, TNext>(
   })
 
   return replicated
+}
+
+export interface ConsumeAsyncIteratorOptions<T, TReturn, TError> {
+  /**
+   * Called on each event
+   */
+  onEvent: (event: T) => void
+  /**
+   * Called once error happens
+   */
+  onError?: (error: TError) => void
+  /**
+   * Called once event iterator is done
+   *
+   * @info If iterator is canceled, `undefined` can be passed on success
+   */
+  onSuccess?: (value: TReturn | undefined) => void
+  /**
+   * Called once after onError or onSuccess
+   *
+   * @info If iterator is canceled, `undefined` can be passed on success
+   */
+  onFinish?: (state: [error: TError, data: undefined, isSuccess: false] | [error: null, data: TReturn | undefined, isSuccess: true]) => void
+}
+
+/**
+ * Consumes an event iterator with lifecycle callbacks
+ *
+ * @warning If no `onError` or `onFinish` is provided, error will be thrown into unhandled rejection channel.
+ * @return unsubscribe callback
+ */
+export function consumeAsyncIterator<T, TReturn, TError = ThrowableError>(
+  iterator: AsyncIterator<T, TReturn> | PromiseWithError<AsyncIterator<T, TReturn>, TError>,
+  options: ConsumeAsyncIteratorOptions<T, TReturn, TError>,
+): () => Promise<void> {
+  void (async () => {
+    let onFinishState: [error: TError, data: undefined, isSuccess: false] | [error: null, data: TReturn | undefined, isSuccess: true]
+
+    try {
+      const resolvedIterator = await iterator
+
+      while (true) {
+        const { done, value } = await resolvedIterator.next()
+
+        if (done) {
+          // if iterator is canceled, value can be undefined
+          const realValue = value as typeof value | undefined
+          onFinishState = [null, realValue, true]
+          options.onSuccess?.(realValue)
+          break
+        }
+
+        options.onEvent(value)
+      }
+    }
+    catch (error) {
+      onFinishState = [error as TError, undefined, false]
+
+      /**
+       * If no `onError` or `onFinish` is provided, unhandled rejections will be thrown
+       * This is best practice for error handling - error should not be silently ignored
+       */
+      if (!options.onError && !options.onFinish) {
+        throw error
+      }
+
+      options.onError?.(error as TError)
+    }
+    finally {
+      options.onFinish?.(onFinishState!)
+    }
+  })()
+
+  return async () => {
+    await (await iterator)?.return?.()
+  }
 }
